# swagger.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/create_currency":{"post":{"tags":["CRUD database"],"summary":"Create Currency","description":"Create currency","operationId":"create_currency_create_currency_post","parameters":[{"required":true,"schema":{"title":"First Currency","type":"string"},"name":"first_currency","in":"query"},{"required":true,"schema":{"title":"Second Currency","type":"string"},"name":"second_currency","in":"query"},{"required":true,"schema":{"title":"Price","type":"integer"},"name":"price","in":"query"},{"required":true,"schema":{"title":"Date","type":"string","format":"date"},"name":"date","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/get_currency/{id}/":{"get":{"tags":["CRUD database"],"summary":"Get Currency","description":"The path parameter for id should have the same name as the argument for id\nso that FastAPI will know that they refer to the same variable\nReturns a currency object if one with the given id exists, else null","operationId":"get_currency_get_currency__id___get","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/list_currency":{"get":{"tags":["CRUD database"],"summary":"List Currency","description":"Fetch a list of all currency object\nReturns a list of objects","operationId":"list_currency_list_currency_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/update_currency/{id}/":{"put":{"tags":["CRUD database"],"summary":"Update Currency","description":"Update currency by id","operationId":"update_currency_update_currency__id___put","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"},{"required":true,"schema":{"title":"Date","type":"string","format":"date"},"name":"date","in":"query"},{"required":true,"schema":{"title":"First Currency","type":"string"},"name":"first_currency","in":"query"},{"required":true,"schema":{"title":"Second Currency","type":"string"},"name":"second_currency","in":"query"},{"required":true,"schema":{"title":"Price","type":"integer"},"name":"price","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete_currency/{id}/":{"delete":{"tags":["CRUD database"],"summary":"Delete Currency","description":"Delete your currency by id","operationId":"delete_currency_delete_currency__id___delete","parameters":[{"required":true,"schema":{"title":"Id","type":"integer"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/get_currency_by_currency_names/{first_currency}/{second_currency}":{"get":{"tags":["CRUD database"],"summary":"Get Currency By Currency Names","description":"The path parameter for currency names should have the same name as the argument for currency names\nso that FastAPI will know that they refer to the same variable\nReturns a currency object if one with the given currency names exists, else null","operationId":"get_currency_by_currency_names_get_currency_by_currency_names__first_currency___second_currency__get","parameters":[{"required":true,"schema":{"title":"First Currency","type":"string"},"name":"first_currency","in":"path"},{"required":true,"schema":{"title":"Second Currency","type":"string"},"name":"second_currency","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/get_currency_price_by_date_and_names/{first_currency}/{second_currency}/{date}":{"get":{"tags":["CRUD database"],"summary":"Get Currency Price By Date And Names","description":"get rate with a names currency on current date","operationId":"get_currency_price_by_date_and_names_get_currency_price_by_date_and_names__first_currency___second_currency___date__get","parameters":[{"required":true,"schema":{"title":"First Currency","type":"string"},"name":"first_currency","in":"path"},{"required":true,"schema":{"title":"Second Currency","type":"string"},"name":"second_currency","in":"path"},{"required":true,"schema":{"title":"Date","type":"string","format":"date"},"name":"date","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/upload":{"post":{"tags":["Files"],"summary":"Upload","description":"Upload your files for db","operationId":"upload_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_upload_post":{"title":"Body_upload_upload_post","required":["input_file"],"type":"object","properties":{"input_file":{"title":"Input File","type":"string","format":"binary"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}